The API now scrapes CoinGecko for a list of coin IDs.
The IDs are saved as values, with the equivalent symbol as the key. There is also a new API endpoint called /coins/, which allows for a GET request with a "symbol" URL query string as a parameter, and returns the ID of a coin based on its symbol. This will later be used to allow users to add holdings using a coin's symbol/ticker rather than its CoinGecko ID. The list of coins is fetched every time the user logs in, refreshes the page, or opens the app, but only if it's been more than an hour since the last time the list was fetched.