The transaction list can now be searched/filtered.
Since searching looks through each transaction's details to determine a match, anything can be searched for. For example, searching for "/19" would return every transaction on the 19th of any month. Searching for "Coinbase" would return every transaction with Coinbase in any of its fields (exchange, from, to and so on).